Implemented _optimize_quantile_count()
_batch_encode() now calls this function by default to find in a more efficient way the
optimal number of quantiles for each source. A new 'score' is defined that finds a
compromise between the competing goals of minimizing epsilon and maximizing the
quantile count.

Changes in density_to_quantiles, binned_to_quantiles and samples_to_quantiles are
made to accept a new keyword target_quantiles.

A few previously implemented functions are added to the API.