Added support for zeta units
New keyword 'units' in encoding, decoding, and plotting functions allows to
specify if the PDF is sampled in units of redshift (z) or zeta = ln(1+z).

Both the API and CLI have been modified to support this feature.

Additionally, some arguments for the encode and decode modes in CLI have been redefined
to accomodate both redshift and zeta units.

A few minor issues in the decode module have been fixed.